Changed data types of shapeId and sequence from long to int & renamed RouteEntityReaderTest to RouteReaderTest & added ShapePointReaderTest for testing ShapePointReader with partial implementation

+++ src/main/java/com/delhitransit/core/model/entity/ShapePointEntity.java
@@ -28,7 +28,7 @@ public class ShapePointEntity {
 
     @Getter
     @Setter
-    private long shapeId;
+    private int shapeId;
 
     @Getter
     @Setter
@@ -40,7 +40,7 @@ public class ShapePointEntity {
 
     @Getter
     @Setter
-    private long sequence;
+    private int sequence;
 
     @Getter
     @Setter
